Add a validator
This new validator check if a parameter in task
config file which accept only few values has the
right value from a list defined.

If the parameter is required the validator check
also if the paremeter is present is the task config file.

If the paremeter is optional, only the right
values are checked.

Change-Id: I1e592317f536c0ab0df90901561c95075ab7f4fa