Added user creds in 'rally deployment check' err debug msg with tests
The authentication error message for 'rally deployment check' was
not very specific. I added username and tenant_name information in
debugging mode to make it easier for users to debug. Notice that the
changes are made in credential layer this time.

Also, I refactored the exceptions handling place by concluding
InvalidEndpointsException and HostUnreachableException into one
AuthenticationFailed exception and moving it from
Clients/verified_keystone into Keystone/auth_ref.
Now, verified_keystone will only check if the user is admin and
all other exceptions will be handled in auth_ref().

For tests, I added two unit tests for the Keystone.auth_ref() and
one functional test to check if the authentication error is caught
and if the corresponding details are printed.

Now, the debugging error msg contains the username & tenant name:
"""
   [-] Unable to connect for user admin in project demo
"""

Change-Id: I3986f883805e95989c7617cea5396f019a4a0211