Fix an issue of database migration
We did not migrate all old atomic to new format, just convert
them while getting data from database and update, so it maybe
occur an error while merging aotmic at line #104. I think it
is the time that we update all old atomic to new format, and
remove the lazy-migration.
NOTE:
It is ok to modify existing migration, since we have 2 cases:
1) the database doesn't include results in old format. In this case the
   old migration succedded and change of it doesn't affect anything.
   Everything is ok here
2) the database includes results in old format. In this case the old
   migration will fail and do not actually apply, so it will be possible
   to apply it again(with new code)

Change-Id: Idb2a0ffd161163537e62ced320f8da033f4e04c9