[ci] Turn off glance scenarios
Glance doesn't work under uswgi. It fails too often(it is really hard to
find the patch where it passed the gates) and we are blocked by it.

This patch changes SLAs, so out gates will continue work, but we will
able to experiment with glance and try to find a solution

Change-Id: I7a45015c1b5ca613499f04b6f0398048e759fd62