[ci] do not fail on glance resources
Glance has an issue with integration with uwsgi. Since OpenStack
community has a global goal - move everyone under uwsgi, glance team
cannot just revert moving under uwsgi, so we need to wait.

Since we do not want to be blocked due to such kind of issues, let's
skip the case when glance resources are not deleted.

Change-Id: I0a8c5d8867f6b43b80a97c2eeda3cefb9f02dd4d