Remove isfile checks on task and task args files
The checks make it impossible to use output redirection (E.g., 'rally
task start --task <(cat *.yaml)') The whole check is unnecessary,
since open(task_file) returns perfectly useful error messages itself,
and works in all the cases we care about -- we don't actually care if
something is a file, only if we can read it.

This also catches errors in case the task or task args files are
unreadable (but are files), which previously would have raised IOError
to the user.

This is a forward-port of ff9c0c1, and fixes the same problem again.

Change-Id: I814fa7aa185d16a6f61baf6ba94cd5e63b00cdfb