Add support for nested atomics for MainStatsTable
Recent refactoring of atomic actions allowed to store nested atomic
actions in the database. There is only one way to see them (using JSON
exporter which is not merged at the time this commit message is
written). It would be nice to not ignore them and expose to users as
much as possible useful information.
This patch adds support of nested atomic actions to the main table of
HTML reports. Since the implementation of it (class MainStatsTable) is
used in several places, this patch makes a bit more useful things:

* The format of storing duration stats in the database.
  Note: the proper database migration is done in the separate patch
        (the last patch in the chain), since the format proposed in this
        patch is not final and it will be changed in the following
        patches. In other words there is no reason to propose the similar
        migrations which will iterate throw all workloads and provide
        temporary result

* Fix the processing of failed iterations. If the iteration has several
  atomic action and one of it fails (not the first one), all the
  previous atomic actions are marked as failed too (but they succeded
  since otherwice all the following actions would not be executed) with
  increasing counters of execution numbers(which is ok), but without
  taking into accaunt the durations (which lead to wrong data of min,
  max,  avg, median, 90%ile, 95%ile stats).

Change-Id: If77bb221c496051a12bd4e5376bae92cecd55668