Fix an issue with missing idle_duration in reports
The total duration of an iteration consists of two components:

 * `duration` - time in seconds if doing something
 * `idle_duration` - time is seconds of sleeping (waiting for something)

In html report (main table) and in duration stats, `duration` was used
as total duration, which lead to strange data in some cases.
This patch fixed the values of total row and adds sub-atomics with
duration and idle_duration (we finally can do it with nested atomics).

Also, since total row is not the last in the table any more, this patch
adds an ability to setup css class name of the rows for html. `total`
row should be bold in any case and all nested atomic actions are italic
now.

Change-Id: I4801fa4de405f9b0dfaaa7ff4e97a9c76b25099a