Fix message for required_params validator
`required_params` validator support the case OR which means that at
least one argument from thelist should be specified. In this case only
the last argument from the list was included in the error message.

This patch fixes an error message in this case. Also, it replaces the
outdated word benchmark by task.

Change-Id: I89dade2f045ad260e70268684fb3f1656fb051d0