Leave only one source of truth for atomic durations
The recent format of atomic actions solved several serios issues which
had been found in the old format:

* ordered dict is not something obvious for the end users, so the
  processing of task results was the pain for new customers (the order
  of atomic actions was lost);

* dict doesn't allow to store several atomics with equal names (the case
  when you need to create several entities of one resource). It lead to
  providing 2 workarounds: auto changing atomic names (in case if the
  name is used, `(i)` suffix had been added to the name); wrapping
  several actions by one parent atomic action like `create_10_resources`
  (which lead to losing timings of the separate actions).

Using list fixes all these issues. Despite the fact that list is quite
good data type which can garantee the similar processing the results in
different languages and tools, not all the places required the extended
information about atomic actions. In combination with inability to port
all places in code to the new atomic action format(it is impossible to
do it at once), it leads to creating temporary dict for charts of HTML
reports. Since this temporaty dict was helpful in a bunch of places, to
reduce creation time of it, we kept it in the database.

After some time this temporary dict become quite similar to the atomic
duration stats which are produced by MainStatsTable. Even more, after
porting MainStatsTable to support nested atomic actions, temporary dict
became a limited version duration stats results which we keep in the
database too.

There is no reason to store temporary dict in the database any more,
since it looks like duplication of duration stats but in the different
formats. To reduce the redundant data and leave only one source of truth
for atomic duration stats, this patch removes usage of temporary dict.

Note: despite the fact that we stopped saving this temporary dict in
      the database for new workloads, there is no database migration in
      current patch. It is done in the separate patch which merges all
      the fixes of workload duration in the one database migration.

Change-Id: I26aac734ef7cf115c767a51d34789e3d05a28241