Update network wrapper to handle neutron plugins router handling
Some plugins (such as Cisco ASR) don't use router ID as the device ID
of the port. Also, some plugins don't allow user to update HA router
interface as it has internal logic to handle that.

This patchset is to catch some exception such that it will not stop
cleanup context to cleanup further resources

Change-Id: Idb1f2922e0310fdeb57cef752d27b7ef11e13865