Fix docstring of VMTasks.dd_load_test scenario
The docstring of VMTasks.dd_load_test missed a lot of details about the
sceanario, which made our customers unhappy with it.

This patch tries to fix the docstring, remove wrong information and add
description of the input arguments.

Also, this patch lists all arguments of VMTasks.dd_load_test explicitly
in the run method.

Change-Id: I5e79c84db223b1be5bd1f3b902f567ad8b2f6ad6