Fix the atomic action table of task results
The atomic action has 2 properties:
 * name - the actual name of the atomic
 * display_name - the user-friendly name

For example, if the atomic action with a name "create_something" is executed
2 times in the one iteration, it will have "create_something (x2)"
display_name.

The was a bug in the main stats table of task detailed results. For the
children atomic actions(nested atomic actions) "dispay_name" was used, but
the top level used "name" property.

Change-Id: I87cc461262657d2c4e032b590df947d576c7638c