Store which atomic fails
The original format of atomic actions did not allow to store much info.
It stored just names and durations, so the iteration data stores the
information about failures.
For the simplification of logic, we made an assumption that the error
had occurred in the last recorded atomic action. This doesn't cover the
case when the error occures in the place which is not covered by
AtomicTimer.
Since we switched to the new format of atomic actions, we can store
explicitly which atomic failed.

For existing data let's leave an assumption that the last recorded
atomic failed.

As for database migration, we already have a unreleased migration which
ports the old atomic actions to the new format, so we can add an
extended migration which will cover 2 cases: port old format of atomic
actions to the new format and mark the last actions as failed. One
migraiont for 2 things can significantly reduce spent time, since both
these actions requires iterating over the all workloads results.

Change-Id: Iebd9645543b9db3e0aba73e3fbd5768d9af3ce2e