Clean docstrings of CLI methods from parameters
There are a lot of commands which have 2 places with descriptions of the
parameters:

 * The first place is docstring of method itself. It is unified
   pythonic way

 * The second one is "help" argument of rally.cli.cliutils.args
   decorator which is designed to initialize argument of a method as CLI
   argument.

It is over-complicated and misleading maintaining the 2 places, so we
need to leave only one source of truth. Since the description from "help"
argument is what user see, let's leave it and restrict describing CLI
methods parameters via regular docstrings.

Change-Id: Idbdf0f68c28ae1bd725506be81c0f8b2cf0a48bc