Modify Bgpvpn scenarios to test true bgpvpn
Currently the bgpvpn scenarios test only a record in the database when
a bgpvpn is created and associated with a network or a router.

To test a true bgpvpn and not only a record in the neutron database
a active port need to be created in the network associated to the bgpvpn.

This change add a server which belong to a network in the context of all
the bgpvpn scenarios concerned. If a bgpvpn driver is used on the openstack,
a true bgpvpn is created.

Change-Id: I64fb6bacf7866332ea55b3c5792bb23db366decd