[atomic actions] Add reset method
ContextManager uses the same instances of contexts (which are inhertors
of ActionTimerMixin) for both setup and cleanup methods. In terms of
storing stats for contexts, it will make cleanup stats include the stats
of setup method as well. It is wrong behavior and it is a case for
extending ActionTimerMixin interface with a method to reset all saved
atomic actions.

Change-Id: I15e9b8aa446c7746fbee1b62e2d62226de4321a8