[openstack] Change identifiers for floating ips
Neutron API of pre-newton OpenStack releases doesn't support any
identifiers for Floating IPs. That is why, to provide a good cleanup we
decided to identify those resources from attached ports.

Actually, it was a wrong decision and in case of existing users case
this feature provides distructive actions.

From OpenStack Newton the "description" field appeared in Floating IP
resource and we can use it as identifier. Such behaviour can help to
remove a bunch of "guessing code".

In case of pre-newton releases, the new config option
(openstack.pre_newton_neutron) should be used to identify that "description"
field should not be added to a resource.

Change-Id: I6b44c09197c0cedd1e11b9dc1b363345051366d4