[env][openstack] Change format of info method
The info method of existing@openstack returns a dict with information
about available services. The key isservice types, the value is
service name. It is convenient for code usage, but not obvious for the
users, as much as "__unknown__" service name.

This patch changes the format to be a list of dicts. Each dict is a
representation of available service. The service type is represented via
"type" field and name (if available) is represented via "name" field.

Also, the new format allows to provide more info about services in
future.

Change-Id: I4be4468d4a78ceb43e1bfa97c35c396606aa4da2