Don't use sitepackages for cli target
It fails on fedora25 with system gssapi package installed with the
following error:

AttributeError: 'module' object has no attribute 'GSSException'

It happens because of: https://github.com/paramiko/paramiko/issues/584
that will be fixed with https://github.com/paramiko/paramiko/pull/861
that is not released or even merged yet.

It's better to isolate the venv from system libraries that may be
incompatible with paramiko.

Change-Id: I98af75bb01fa42a2902a850de1b974abc982d432