[task] Move TaskConfig to separate module
The validation and processing of input task config to the unified
version-aware object is done by special class TaskConfig. It supports
both v1 and v2 formats. The size of this helper class is quite big and
since we are planning to extend it, it is reasonable to move it to
separate module - rally.task.task_cfg .

NOTE: we do not need to deprecate the old place, since TaskConfig is our
inner class and should not be used outside our code base.

Change-Id: I875b677a689e002413d0f07d43bacc960e965df4