[verify] Ger rid of calling pip internals
In case of system-wide mode of verifiers, Rally tryies to check whether
tool's requirements feets what we have installed in the system
environment or not.

Previously we used helpers from pip library. Unfortunately, pip team
decided to mark these helpers as "internal" and it becomes an issue for
us with pip>=10.0

This patch migrates our checks to use pkg_resources .

Change-Id: I6f5fb6c1f71c3e0b2c59481b00907d08e32c8eb1