[types] Remove redundant function
Function `rally.task.convert` is designed to specify which arguments of
scenario plugin should be preprocessed. It accepts keyword arguments
where key is a argument name and value is options for processing.

This function also includes a logic for postponing the loading of
ResourceType plugins to the place where they are actually executed.

I do not know historical reasons, but as for currect code place, we just
can call for obtaining the right ResourceType plugin in the right place
without any wrappers and so on.

Change-Id: Ide57d48301eba7022f2365b2d68615e5431a4c11