Redesign ResourceTypes
This patch redesigns base class of ResourceType's so it can handle:
* non-openstack related resources
* cache mechanism

Plugins of `ResourceType` type look like the last place hardcoded for
OpenStack. It's base method "transform" expects "clients" argument with
rally.plugins.openstack.osclients.Clients instance. This instance is
initialized in our base code at rally.task.types.

To solve this issue, the new instance method "pre_process" is
introduced. It expects the resource specification which can be used for
searching/filtering + a configuration for ResourceType from the plugin.

Using instancemethod instead of classmethods allows to share utils and
simplify introducing cache-feature which can decrease number of equal
calls.

Despite the fact, that the design of ResourceType class is complitely
changed, everything is done in a backward compatible way.

Change-Id: Iea95450892f229886c86697d632013e868b6cb31