Remove hardcode to openstack from required_platform validator
This validator is one of the general one. Despite the fact, it is not
directly hardcoded to OpenStack plaform, it's interface was desgined to
openstack credentials and usage of it for non-openstack platforms looks
ugly.

This patch changes it's behaviour to check existance of the platform in
the environemnt spec for main case + searchs&re-calls
required_platfrom@platform_name for custom logic like checking
admin/user credentials.

Change-Id: I546bbd9b0e35abb30195b0f8c98756a9cce10107