Task context: implement comparison operators
__cmp__ is no longer used in python3

Other operators need to be defined instead.

https://docs.python.org/3/whatsnew/3.0.html#ordering-comparisons
"The cmp() function should be treated as gone, and the __cmp__() special
method is no longer supported. Use __lt__() for sorting, __eq__() with
__hash__(), and other rich comparisons as needed. (If you really need
the cmp() functionality, you could use the expression (a > b) - (a < b)
as the equivalent for cmp(a, b).)"

https://www.python.org/dev/peps/pep-0207/

This change does it based on get_order and class meta properties.

Downstream projects that rely on context comparison (such as
rally-openstack) will benefit from this change.

Change-Id: I4991acccefd4754e209feafd5e24d03c76c283f8
Depends-On: I42db6045ae88c34d3576beb737bb7e955c5f2132
Needed-By: I6bb31fe1d3f79a8fd4707a5576204dee69a3f307
Partial-Bug: #1819274